feat: Add Loader component for displaying loading state
This commit adds a new Loader component to the application. The Loader component is responsible for displaying a loading state to the user while data is being fetched or processed. It includes a spinning loader animation and a text indicating that the page is loading. This component will be used throughout the application to improve the user experience.